Fixed hole size on trimmer cap
Fixed hole size on trimmer cap in library PCA9536.lbr and updated the
two autoband boards to use the corrected package. Package can be
replaced on other boards with the one in this library to fix the
problem. Also moved the text on the horizontal board to the bPlace
layer. Couldn't do this on the smaller board as no where to put it!